backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
# Assuming torch is already installed with CUDA support as per user context, 
# but listing it here for completeness.
# torch 
# onnxruntime-gpu
requests
tqdm
jinja2
python-jose[cryptography]
passlib[bcrypt]
